Use a fixed number type
The motivation for using fixed number types is to make the scripting
language easier to use for writers. Instead of having to explain the
intricacies of floating point numbers, it is much easier to have just
one restriction: The number has a maximum precision after the decimal
point.

When writing narrative dialog, picking a high enough precision makes it
such that the number allows you to do most things you'd want to do in a
narrative language like keep track of currency values or do basic
arithmetic without having to worry about how the number is actually
implemented in the computer.