Initial node parsing
Adds initial parsing for say and append nodes. These nodes are
implemented a little differently from substitutions which require quotes
to surround the text. Instead, no quotes are needed since these are
difficult to remember and explain to writers. This also has the added
benefit of making the code read more like a screenplay.

However, quoted strings are still required when specifying strings in
substitutions. These are required, otherwise it would not be possible to
distinguish them from variables.

As a side effect of these changes, the text simplification has been
updated to always strip leading and trailing whitespace from text.