Add Choice node
I intend to add the ability to give choices an identifier so they can
be referenced later (such as in the case of specifying a default
choice).

However, the optional identifier needs some way of being distinguished
from the rest of the script. I have decided to surround the identifier
in brackets. To make this more consistent in the language, I have also
updated the label parsing to use the same syntax.