Check maximum size of Set
According to Data.Set documentation, the size of the set must not exceed
`maxBound::Int`. Violation of this condition is not detected and if the
size limit is exceeded, its behaviour is undefined.

This means that there is an upper bound on how many labels we can
define, and although it's unlikely any user will ever hit this bound it
makes more sense for the parser to fail immediately rather than let
the program continue.