Add Script type
This data type will represent the object form of a parsed Rumor script.

Note that while `Section` is currently a constructor in the `Node`
sumtype, we would like to keep them separate from normal nodes to
simplify the execution of a script in the future. This is why the
`sections` record exists on the `Script` data type.