Add StackFrame type
The stack frame will be used to keep track of where we are when we are
executing a rumor script. Note that I've opted to remove the use of the
`Integer` type since the `Safe` module only uses `Int`s and it would
be cumbersome and not terribly beneficial to continue using the
`Integer` type for now.