Automatically generate labels for choices
Instead of relying on an index to refer to a choice a user would like to
pick when an identifier isn't supplied, it's easier to automatically
generate an identifier to refer to that choice with. This avoids
the ambiguity of an index, which may or may not still point to the
correct choice.

This label is generated by making a SHA1 hash of 120 random Word8
characters generated by a random number generator seeded with the
hash of the script. This allows the hashes to be exactly the same if the
script isn't changed and recompiled.