Use user state to record sections
We need to add block support to the `choice` node, but it differs from
either of the two existing cases as it is both a node and contains a new
section. It would be really cumbersome to update the `choice` parser to
return both a Node and a map of sections, so i've opted to store the
sections in the user state instead.

This also removes the need for the `OneOf` type.