Add initial FFI interface
By default, the entry points of all the object files will be exported
from the DLL when using the `-shared` flag, which causes us to exceed
the maximum symbol count for DLLs in Windows. This is fixed by defining
a `.def` file to explicitly list what we would like to export.
( see: https://gitlab.haskell.org/ghc/ghc/-/issues/18030 )

I decided to use a `StablePtr (Script E12)` as the return type for
`parse` because it's not trivial to write a `Storable` instance due to
the fact that nodes are implemented as sumtypes, which have no
corresponding type in C.

I also defaulted to the `E12` precision for the fixed number type just
to make it easier for now to implement the FFI. Ideally in the future,
we would somehow let the caller decide the precision of the fixed number
type.

For some reason, this causes the internal tests to fail, but I plan on
removing them anyway, so I've decided to not investigate further.