Fix segmentation fault on hs_init
GHC does not mangle any of the names we export in our FFI, so the `free`
function that was defined to free stable pointers effectively overwrote
the already-existing `free` function from C that GHC was expecting to
call. This caused a segmentation fault when calling hs_init from foreign
code.

Instructions for how to build a shared library on Windows and some
documentation has been added summarizing some of the relevant things I
have learned figuring this bug out with the help of a friend.